+++ IntRX/Main.py
@@ -1,17 +1,9 @@
 from Interact import *
-import datetime
-import re
-import time
-import json
-from websocket import create_connection
 from threading import Thread
 from win32gui import GetWindowText, GetForegroundWindow
 
 
-interact = InteractGame()
-currentCommands = None
-
-
+# TODO - REWORK THIS. Split the phrase to a left and right str, check if left is in cmd and right is in cmd, and if so remove both and you're left with the cmd.
 class cmdPhrase:  # Anticommands, my name for grabbing the command from a chat message rather than a command.
     def __init__(self):
         self.fromLeft = 0
@@ -45,116 +37,120 @@ class cmdPhrase:  # Anticommands, my name for grabbing the command from a chat m
         elif self.fromRight:
             toreturn = cmd[:-self.fromRight]
 
-        print(toreturn)
-        return toreturn.replace("!", '')
+        return toreturn.replace(settings["PREFIX"], '')
 
 
 def formatted_time():
     return datetime.datetime.today().now().strftime("%I:%M")
 
 
-def runCmdExtras(command, cmdarguments, item):
-    tempCooldown = datetime.datetime.now() + datetime.timedelta(seconds=item[1])
-    tempGlobalCooldown = datetime.datetime.now() + datetime.timedelta(seconds=settings["CD BETWEEN CMDS"])
-    cooldowns.update({command: tempCooldown})
-    cooldowns.update({"00rx_globalCD": tempGlobalCooldown})
-    writeArgs(cmdarguments)
-    print(item[0] + " executed!")
 
 
-def globalCooldown():
-    global cooldowns
-    if "00rx_globalCD" in cooldowns.keys():  # This is just weird so theres no way anyone will ever accidentally have their command say this. Note, if your command says this and you ask for help, ill be pissed.
-        timeleft = cooldowns["00rx_globalCD"] - datetime.datetime.now()
-        timeleftSecs = round(timeleft.total_seconds())
-        return timeleftSecs
-    return 0
 
 
-def cmdCooldown(command):
-    global cooldowns
-    if command in cooldowns.keys():
-        timeleft = cooldowns[command] - datetime.datetime.now()
-        timeleftSecs = round(timeleft.total_seconds())
-        return timeleftSecs
-    return 0
+
+class commands:
+    def __init__(self):
+        self.cooldowns = {}
+        self.loadedGameCommands = []
+        self.globalCommands = importGlobal()
+        self.activeGame = None
+
+    def runcommand(self, command, cmdArguments, user):
+        run = False
+
+        for item in self.loadedGameCommands, self.globalCommands:
+            for i in item:
+                if command.lower() == i[0].lower():
+                    run = True
 
 
-def runcommand(command, cmdArguments, user):
-    global currentCommands, activeGame, cooldowns, globalCommands
-    if not currentCommands:
-        currentCommands = []
-    run = False
+        if not run:
+            chatConnection.sendToChat("Invalid command")
 
-    for item in currentCommands, globalCommands:
-        for i in item:
-            if command.lower() == i[0].lower():
-                run = True
-
-
-    if not run:
-        chatConnection.sendToChat("Invalid command")
-
-    if run:
-        # Manage cooldowns
-        if globalCooldown() > cmdCooldown(command):
-            chatConnection.sendToChat("Since a command was run recently, nobody can interact for %s more seconds." % globalCooldown())
-            return
-        elif cmdCooldown(command) > globalCooldown():
-            chatConnection.sendToChat("That command is still on cooldown for %s seconds." % cmdCooldown(command))
-            return
-
-        for item in currentCommands:  # Test if the command run is for a loaded game
-            if command.lower() == item[0].lower():  # Command detected, pass this to the InteractGame class.
-                cmdToRun = item[2]
-                cooldown = item[1]
-                if "%ARGS" in cmdToRun and not cmdArguments:
-                    chatConnection.sendToChat("That command requires you to provide an argument to run.")
-                    return
-                cmdToRun = cmdToRun.replace("%ARGS%", cmdArguments)
-                cmdToRun = cmdToRun.replace("%USER%", user)
-                runCmdExtras(command, cmdArguments, item)
-                interact(activeGame, cmdToRun, cooldown, cmdArguments, user)
+        if run:
+            # Manage cooldowns
+            if self.globalCooldown() > self.cmdCooldown(command):
+                chatConnection.sendToChat("Since a command was run recently, nobody can interact for %s more seconds." % self.globalCooldown())
+                return
+            elif self.cmdCooldown(command) > self.globalCooldown():
+                chatConnection.sendToChat("That command is still on cooldown for %s seconds." % self.cmdCooldown(command))
                 return
 
-        for item in globalCommands:  # Test if command run is a global command
-            subOnly = item[4]
-            donoReq = item[5]
-            reward = item[6]
-            if command.lower() == item[0].lower():  # Command detected, run the file
-
-                #if subOnly:
-
-
-
-                if donoReq:
-                    if not chatConnection.donoAmt > donoReq:
-                        chatConnection.sendToChat("This command can only be run with a donation of at least $" + str(donoReq))
-                        return
-
-                if item[2][0] == "$":  # Process built-in global script
-                    if isValidInt(cmdArguments):  # Process MAX ARG setting
-                        if int(cmdArguments) >= settings["MAX ARG"]:
-                            chatConnection.sendToChat("That value is too high, please try again with a lower number.")
-                            return
-
-                    if not processBuiltInGlobal(item[2], cmdArguments, user):  # This runs the exe
+            for item in cmd.loadedGameCommands:  # Test if the command run is for a loaded game
+                if command.lower() == item[0].lower():  # Command detected, pass this to the InteractGame class.
+                    cmdToRun = item[2]
+                    cooldown = item[1]
+                    if "%ARGS" in cmdToRun and not cmdArguments:
                         chatConnection.sendToChat("That command requires you to provide an argument to run.")
                         return
-
-                    runCmdExtras(command, cmdArguments, item)
+                    cmdToRun = cmdToRun.replace("%ARGS%", cmdArguments)
+                    cmdToRun = cmdToRun.replace("%USER%", user)
+                    self.runCmdExtras(command, cmdArguments, item)
+                    interact(cmd.activeGame, cmdToRun, cooldown, cmdArguments, user)
                     return
 
-                if item[3]:  # Do this stuff if there's a specified active window
-                    if item[3] in GetWindowText(GetForegroundWindow()):
-                        runCmdExtras(command, cmdArguments, item)
+            for item in self.globalCommands:  # Test if command run is a global command
+                subOnly = item[4]
+                donoReq = item[5]
+                reward = item[6]
+                if command.lower() == item[0].lower():  # Command detected, run the file
+
+                    # RUN PRE-COMMAND CHECKS to make sure the user can actually run it
+                    if subOnly:
+                        if not mainChatConnection.isSubscriber:
+                            chatConnection.sendToChat("This command can only be run by subscribers.")
+                            return
+
+                    if donoReq:
+                        if not mainChatConnection.bitsAmount > donoReq:
+                            chatConnection.sendToChat("This command can only be run with a donation of at least " + str(donoReq))
+                            return
+
+
+                    if item[2][0] == "$":  # Process built-in global script
+                        if isValidInt(cmdArguments):  # Process MAX ARG setting
+                            if int(cmdArguments) >= settings["MAX ARG"]:
+                                chatConnection.sendToChat("That value is too high, please try again with a lower number.")
+                                return
+
+                        if not processBuiltInGlobal(item[2], cmdArguments, user):  # This runs the exe
+                            chatConnection.sendToChat("That command requires you to provide an argument to run.")
+                            return
+
+                        self.runCmdExtras(command, cmdArguments, item)
+                        return
+
+                    if item[3]:  # Do this stuff if there's a specified active window
+                        if item[3] in GetWindowText(GetForegroundWindow()):
+                            self.runCmdExtras(command, cmdArguments, item)
+                            script.runAHK(r"..\Config\UserScripts\\" + item[2])
+                            return
+
+                    else:  # Otherwise just active the command
+                        self.runCmdExtras(command, cmdArguments, item)
                         script.runAHK(r"..\Config\UserScripts\\" + item[2])
                         return
 
-                else:  # Otherwise just active the command
-                    runCmdExtras(command, cmdArguments, item)
-                    script.runAHK(r"..\Config\UserScripts\\" + item[2])
-                    return
+
+    def globalCooldown(self):
+        if "00rx_globalCD" in self.cooldowns.keys():  # This is just weird so theres no way anyone will ever accidentally have their command say this. Note, if your command says this and you ask for help, ill be pissed.
+            timeleft = self.cooldowns["00rx_globalCD"] - datetime.datetime.now()
+            return round(timeleft.total_seconds())
+        return 0
+
+
+    def cmdCooldown(self, command):
+        if command in self.cooldowns.keys():
+            timeleft = self.cooldowns[command] - datetime.datetime.now()
+            return round(timeleft.total_seconds())
+        return 0
+
+    def runCmdExtras(self, command, cmdarguments, item):
+        self.cooldowns.update({command: datetime.datetime.now() + datetime.timedelta(seconds=item[1])})
+        self.cooldowns.update({"00rx_globalCD": datetime.datetime.now() + datetime.timedelta(seconds=settings["CD BETWEEN CMDS"])})
+        writeArgs(cmdarguments)
+        print(item[0] + " executed!")
 
 
 class mainChat:
@@ -164,12 +160,12 @@ class mainChat:
         self.rewardTitle = None
         self.rewardCost = None
         self.isSubscriber = False
-        global settings, globalCommands
-        globalCommands = importGlobal()
+        global settings
 
     def main(self):
         chatConnection.start()
         while True:
+            prefix = settings["PREFIX"]
             result = chatConnection.ws.recv()
             resultDict = json.loads(result)
             print(resultDict)
@@ -202,6 +198,7 @@ class mainChat:
                         subLevel = resultDict["event"]["sub_level"]
                         print(
                             "(" + misc.formatTime() + ")>> [EVENT] " + subUsername + " subscribed with level %s for %s months." % (subLevel, subMonths))
+                        self.isSubscriber = True
                     except:
                         pass
 
@@ -220,20 +217,24 @@ class mainChat:
                             user = resultDict["event"]["sender"]["displayname"]
                             command = ((message.split(' ', 1)[0]).lower()).replace("\r", "")
                             cmdarguments = message.replace(command or "\r" or "\n", "")[1:]
+                            if "SUBSCRIBER" in resultDict["event"]["sender"]["roles"]:
+                                self.isSubscriber = True
+                            else:
+                                self.isSubscriber = False
                             print("(" + misc.formatTime() + ")>> " + user + ": " + message)
 
 
                             if settings["COMMAND PHRASE"]:  # Process anticommands
-                                if user in [settings["BOT NAME"], settings["ALT BOT NAME"]]:
+                                if user.lower() in [settings["BOT NAME"].lower(), settings["ALT BOT NAME"].lower()]:
                                     if commandPhrase.trimSetting() in message.lower():
                                         extractedCmd = commandPhrase.extractCmd(message)
                                         cmdarguments = (extractedCmd.replace(extractedCmd.split(" ")[0], '')).strip()
                                         toRun = extractedCmd.split(" ")[0]
                                         print("Running command from another bot: " + toRun)
-                                        runcommand("!" + toRun, cmdarguments, user)
+                                        cmd.runcommand(prefix + toRun, cmdarguments, user)
 
-                            elif command[0] == "!":  # Only run normal commands if COMMAND PHRASE is blank
-                                runcommand(command, cmdarguments, user)
+                            elif command[0] == prefix:  # Only run normal commands if COMMAND PHRASE is blank
+                                cmd.runcommand(command, cmdarguments, user)
 
                             self.bitsAmount = 0
 
@@ -267,55 +268,54 @@ supportedGames = {
     "Fallout3": "Fallout 3",
     "Minecraft": "Minecraft",
     "Subnautica": "Subnautica",
+    "Subnautica Below Zero": "Subnautica",
     "The Witcher 3": "Witcher 3",
     "Valheim": "Valheim"
 }
 
 
 def refresh():
-    global currentCommands, activeGame
     cacheActiveGame = ''
-    activeGame = None
+    cmd.activeGame = None
     while True:
-        time.sleep(int(settings['REFRESH INTERVAL']))
+        time.sleep(2)
         currentWindow = GetWindowText(GetForegroundWindow())
 
-        activeGame = None
+        cmd.activeGame = None
 
         for winName in supportedGames:
-            if winName in currentWindow and (activeGame != supportedGames[winName]):
-                activeGame = supportedGames[winName]
+            if winName in currentWindow and (cmd.activeGame != supportedGames[winName]):
+                cmd.activeGame = supportedGames[winName]
 
-        if cacheActiveGame != activeGame and activeGame:  # Do when user starts NEW game
-            print("Now playing " + activeGame)
-            cacheActiveGame = activeGame
-            if settings['ANNOUNCE GAME'].lower():
-                chatConnection.sendToChat("The streamer is now playing " + activeGame + " and you can interact with it!")
-            currentCommands = importInteraction(activeGame)
+        if cacheActiveGame != cmd.activeGame and cmd.activeGame:  # Do when user starts NEW game
+            print("Now playing " + cmd.activeGame)
+            cacheActiveGame = cmd.activeGame
+            if settings['ANNOUNCE GAME']:
+                chatConnection.sendToChat("The streamer is now playing " + cmd.activeGame + " and you can interact with it!")
+            cmd.currentCommands = importInteraction(cmd.activeGame)
 
-        elif not activeGame:  # If no game is loaded
-            currentCommands = []
+        elif not cmd.activeGame:  # If no game is loaded
+            cmd.currentCommands = []
 
-        elif activeGame and (len(currentCommands) == 0):  # If the game is tabbed back in, but not a new game (don't announce it)
-            currentCommands = importInteraction(activeGame)
+        elif cmd.activeGame and (len(cmd.loadedGameCommands) == 0):  # If the game is tabbed back in, but not a new game (don't announce it)
+            cmd.currentCommands = importInteraction(cmd.activeGame)
 
 
 def tick():
-    global cooldowns
-    cooldowns = {}
-    while True:
+    while True:  # Test for cooldowns being complete
         time.sleep(0.5)
-        if cooldowns:
-            for item in cooldowns:
-                if cooldowns[item] <= datetime.datetime.now():
-                    cooldowns.pop(item)
+        if cmd.cooldowns:
+            for item in cmd.cooldowns:
+                if cmd.cooldowns[item] <= datetime.datetime.now():
+                    cmd.cooldowns.pop(item)
                     break
 
 
 mainChatConnection = mainChat()
 commandPhrase = cmdPhrase()
+cmd = commands()
 
-t1 = Thread(target=mainChatConnection.main())
+t1 = Thread(target=mainChatConnection.main)
 t2 = Thread(target=refresh)
 t3 = Thread(target=tick)